+++ whetstone/compiler/src/main/kotlin/com/freeletics/mad/whetstone/codegen/util/External.kt
@@ -29,7 +29,7 @@ internal val asComposeState = MemberName("com.freeletics.mad.whetstone.internal"
 internal val internalWhetstoneApi = ClassName("com.freeletics.mad.whetstone.internal", "InternalWhetstoneApi")
 internal val navEntryViewModelProvider = MemberName("com.freeletics.mad.whetstone.internal", "viewModelProvider")
 internal val fragmentViewModelProvider = MemberName("com.freeletics.mad.whetstone.fragment.internal", "viewModelProvider")
-internal val rememberViewModelProvider = MemberName("com.freeletics.mad.whetstone.internal", "rememberViewModelProvider")
+internal val rememberViewModelProvider = MemberName("com.freeletics.mad.whetstone.compose.internal", "rememberViewModelProvider")
 internal val navEntryComponentGetter = ClassName("com.freeletics.mad.whetstone.internal", "NavEntryComponentGetter")
 internal val navEntryComponentGetterKey = ClassName("com.freeletics.mad.whetstone.internal", "NavEntryComponentGetterKey")
 internal val composeProviderValueModule = ClassName("com.freeletics.mad.whetstone.internal", "ComposeProviderValueModule")
