+++ whetstone/runtime/build.gradle
@@ -60,7 +60,5 @@ dependencies {
     api "androidx.navigation:navigation-common:2.4.1"
     api project(":state-machine")
 
-    implementation "androidx.compose.ui:ui:1.1.1"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.4.1"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core-jvm:1.6.0"
 }
